♻️ 081TPS context: decompose ingest pipeline
