+++ .agentplane/tasks/202605290018-081TPS/README.md
@@ -1,0 +1,122 @@
+---
+id: "202605290018-081TPS"
+title: "Context ingest pipeline decomposition"
+status: "DOING"
+priority: "med"
+owner: "CODER"
+revision: 4
+origin:
+  system: "manual"
+depends_on: []
+tags:
+  - "code"
+  - "hotspot"
+  - "refactor"
+verify: []
+plan_approval:
+  state: "approved"
+  updated_at: "2026-05-29T00:19:03.124Z"
+  updated_by: "ORCHESTRATOR"
+  note: null
+verification:
+  state: "pending"
+  updated_at: null
+  updated_by: null
+  note: null
+  attempts: 0
+commit: null
+comments:
+  -
+    author: "CODER"
+    body: "Start: Decompose context ingest manifest and source collection helpers while preserving cmdContextIngest behavior."
+events:
+  -
+    type: "status"
+    at: "2026-05-29T00:19:20.810Z"
+    author: "CODER"
+    from: "TODO"
+    to: "DOING"
+    note: "Start: Decompose context ingest manifest and source collection helpers while preserving cmdContextIngest behavior."
+doc_version: 3
+doc_updated_at: "2026-05-29T00:19:20.810Z"
+doc_updated_by: "CODER"
+description: "Decompose packages/agentplane/src/context/ingest.ts into focused ingestion pipeline modules while preserving context ingest behavior and reducing runtime hotspot warnings."
+sections:
+  Summary: |-
+    Context ingest pipeline decomposition
+
+    Decompose packages/agentplane/src/context/ingest.ts into focused ingestion pipeline modules while preserving context ingest behavior and reducing runtime hotspot warnings.
+  Scope: |-
+    - In scope: Decompose packages/agentplane/src/context/ingest.ts into focused ingestion pipeline modules while preserving context ingest behavior and reducing runtime hotspot warnings.
+    - Out of scope: unrelated refactors not required for "Context ingest pipeline decomposition".
+  Plan: |-
+    Plan:
+    1. Start branch_pr worktree for CODER.
+    2. Extract context ingest manifest/model helpers and source collection helpers from packages/agentplane/src/context/ingest.ts into focused modules.
+    3. Keep cmdContextIngest as the orchestration facade and preserve existing public command types.
+    4. Verify with focused context ingest tests, typecheck, arch/knip/lint/format, and hotspot threshold report.
+    5. Open PR, wait for hosted checks, merge to main, close lifecycle, and clean the worktree.
+
+    Acceptance:
+    - context ingest behavior remains compatible.
+    - ingest.ts drops below the 400-line hotspot warning threshold.
+    - runtime hotspot warning count decreases from 31 to 30 without adding new warning-sized runtime modules.
+  Verify Steps: |-
+    PLANNER fallback scaffold. Replace with task-specific acceptance checks when PLANNER context is available.
+
+    1. Review the changed artifact or behavior for the `code` task. Expected: the requested outcome is visible and matches the approved scope.
+    2. Run the most relevant validation step for the `code` task. Expected: it succeeds without unexpected regressions in touched scope.
+    3. Compare the final result against the task summary and scope. Expected: any remaining follow-up is explicit in ## Findings.
+  Verification: |-
+    <!-- BEGIN VERIFICATION RESULTS -->
+    <!-- END VERIFICATION RESULTS -->
+  Rollback Plan: |-
+    - Revert task-related commit(s).
+    - Re-run required checks to confirm rollback safety.
+  Findings: ""
+id_source: "generated"
+---
+## Summary
+
+Context ingest pipeline decomposition
+
+Decompose packages/agentplane/src/context/ingest.ts into focused ingestion pipeline modules while preserving context ingest behavior and reducing runtime hotspot warnings.
+
+## Scope
+
+- In scope: Decompose packages/agentplane/src/context/ingest.ts into focused ingestion pipeline modules while preserving context ingest behavior and reducing runtime hotspot warnings.
+- Out of scope: unrelated refactors not required for "Context ingest pipeline decomposition".
+
+## Plan
+
+Plan:
+1. Start branch_pr worktree for CODER.
+2. Extract context ingest manifest/model helpers and source collection helpers from packages/agentplane/src/context/ingest.ts into focused modules.
+3. Keep cmdContextIngest as the orchestration facade and preserve existing public command types.
+4. Verify with focused context ingest tests, typecheck, arch/knip/lint/format, and hotspot threshold report.
+5. Open PR, wait for hosted checks, merge to main, close lifecycle, and clean the worktree.
+
+Acceptance:
+- context ingest behavior remains compatible.
+- ingest.ts drops below the 400-line hotspot warning threshold.
+- runtime hotspot warning count decreases from 31 to 30 without adding new warning-sized runtime modules.
+
+## Verify Steps
+
+PLANNER fallback scaffold. Replace with task-specific acceptance checks when PLANNER context is available.
+
+1. Review the changed artifact or behavior for the `code` task. Expected: the requested outcome is visible and matches the approved scope.
+2. Run the most relevant validation step for the `code` task. Expected: it succeeds without unexpected regressions in touched scope.
+3. Compare the final result against the task summary and scope. Expected: any remaining follow-up is explicit in ## Findings.
+
+## Verification
+
+<!-- BEGIN VERIFICATION RESULTS -->
+<!-- END VERIFICATION RESULTS -->
+
+## Rollback Plan
+
+- Revert task-related commit(s).
+- Re-run required checks to confirm rollback safety.
+
+## Findings
