+++ .agentplane/tasks/202605290018-081TPS/pr/github-body.md
@@ -1,0 +1,33 @@
+Task: `202605290018-081TPS`
+Title: Context ingest pipeline decomposition
+Canonical task record: `.agentplane/tasks/202605290018-081TPS/README.md`
+
+## Summary
+
+Context ingest pipeline decomposition
+
+Decompose packages/agentplane/src/context/ingest.ts into focused ingestion pipeline modules while preserving context ingest behavior and reducing runtime hotspot warnings.
+
+## Scope
+
+- In scope: Decompose packages/agentplane/src/context/ingest.ts into focused ingestion pipeline modules while preserving context ingest behavior and reducing runtime hotspot warnings.
+- Out of scope: unrelated refactors not required for "Context ingest pipeline decomposition".
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T00:19:20.924Z
+- Branch: task/202605290018-081TPS/context-ingest-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
