+++ .agentplane/tasks/202605290018-081TPS/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 081TPS task: Context ingest pipeline decomposition [202605290018-081TPS]
